sdks/fs-python/requirements.txt
pulsar-client>=3.0.0
pyyaml>=6.0
aiohttp>=3.8.0
pydantic>=2.0.0
pytest-asyncio>=1.0.0